+++ audiomate/corpus/base.py
@@ -184,6 +184,26 @@ class CorpusView(metaclass=abc.ABCMeta):
 
         return duration
 
+    def all_tokens(self, delimiter=' ', label_list_ids=None):
+        """
+        Return a list of all tokens occurring in one of the labels in the corpus.
+
+        Args:
+            delimiter (str): The delimiter used to split labels into tokens
+                             (see :meth:`audiomate.corpus.assets.Label.tokenized`).
+            label_list_ids (list): If not None, only labels from label-lists with an idx contained in this list
+                                   are considered.
+
+        Returns:
+             :class:`set`: A set of distinct tokens.
+        """
+        tokens = set()
+
+        for utterance in self.utterances.values():
+            tokens = tokens.union(utterance.all_tokens(delimiter=delimiter, label_list_ids=label_list_ids))
+
+        return tokens
+
     #
     #   Data
     #